// src/app/interfaces/IVentas.ts
export class IVenta {
    idVenta?: number|any;
    idCliente?: number|any;
    totalProducto?: number;
    total?: number;
    contacto?: string;
    idLocalidad?: string;
    telefono?: string;
    direccion?: string;
    idTransaccion?: string;
    fechaVenta?: string; // Usaremos una cadena para representar la fecha
  }
  